+++ apps/ymir-sdl3/src/app/ui/views/settings/video_settings_view.cpp
@@ -75,6 +75,8 @@ void VideoSettingsView::Display() {
     }
 
     MakeDirty(ImGui::Checkbox("Double-click to toggle full screen", &settings.doubleClickToFullScreen));
+    widgets::ExplanationTooltip("This option will not work if you are using a Virtua Gun or Shuttle Mouse.",
+                                m_context.displayScale);
 
     ImGui::Separator();
 
@@ -86,7 +88,7 @@ void VideoSettingsView::Display() {
 
     MakeDirty(ImGui::Checkbox("Synchronize video in full screen mode", &settings.syncInFullscreenMode));
     widgets::ExplanationTooltip(
-        "When enabled, synchronizes GUI updates with emulator rendering while in full screenm mode.\n"
+        "When enabled, synchronizes GUI updates with emulator rendering while in full screen mode.\n"
         "This greatly improves frame pacing but may reduce GUI performance.",
         m_context.displayScale);