+++ apps/ymir-sdl3/src/app/ui/windows/peripheral_config_window.cpp
@@ -12,7 +12,8 @@ PeripheralConfigWindow::PeripheralConfigWindow(SharedContext &context)
     , m_analogPadView(context)
     , m_arcadeRacerView(context)
     , m_missionStickView(context)
-    , m_virtuaGunView(context) {}
+    , m_virtuaGunView(context)
+    , m_shuttleMouseView(context) {}
 
 void PeripheralConfigWindow::Open(uint32 portIndex, uint32 slotIndex) {
     m_portIndex = std::min(portIndex, 1u);
@@ -54,6 +55,7 @@ void PeripheralConfigWindow::DrawContents() {
     case ArcadeRacer: m_arcadeRacerView.Display(settings.arcadeRacer, m_portIndex); break;
     case MissionStick: m_missionStickView.Display(settings.missionStick, m_portIndex); break;
     case VirtuaGun: m_virtuaGunView.Display(settings.virtuaGun, m_portIndex); break;
+    case ShuttleMouse: m_shuttleMouseView.Display(settings.shuttleMouse, m_portIndex); break;
     }
 }