+++ apps/ymir-sdl3/src/app/ui/windows/peripheral_config_window.hpp
@@ -6,6 +6,7 @@
 #include <app/ui/views/settings/arcade_racer_config_view.hpp>
 #include <app/ui/views/settings/control_pad_config_view.hpp>
 #include <app/ui/views/settings/mission_stick_config_view.hpp>
+#include <app/ui/views/settings/shuttle_mouse_config_view.hpp>
 #include <app/ui/views/settings/virtua_gun_config_view.hpp>
 
 namespace app::ui {
@@ -29,6 +30,7 @@ private:
     ArcadeRacerConfigView m_arcadeRacerView;
     MissionStickConfigView m_missionStickView;
     VirtuaGunConfigView m_virtuaGunView;
+    ShuttleMouseConfigView m_shuttleMouseView;
 };
 
 } // namespace app::ui