+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Introduced save state file version 12.
 - Input: Implemented two mouse capture modes:
     - System mouse: binds the system mouse cursor to a single peripheral. Mouse cursor is still available to interact with the GUI.
     - Physical mouse: binds one or more mice to different peripherals. Disables the system cursor while any mice is bound.
+- Input: Implemented Shuttle Mouse peripheral. (#32)
 - Input: Implemented Virtua Gun peripheral. (#33)
 - MIDI: Force RtMidi to use dummy API if it fails to initialize, allowing Ymir to run without MIDI drivers.
 - System: You can now select a preferred system variant (Saturn, HiSaturn, V-Saturn or Dev Kit) and Ymir will automatically pick a matching IPL ROM. (#637, #725; @Fueziwa)
@@ -28,6 +29,7 @@ Introduced save state file version 12.
 - VDP1: Properly load save state data when threaded VDP1 rendering is enabled.
 - VDP1: Stall VDP1 drawing on VRAM writes exclusively on Mega Man X3 and Rockman X3 to fix garbled sprites. (#244)
 - VDP2: Consolidate sprite data handling and fix 16-bit readout of 8-bit sprite data. Fixes garbled graphics in NBA Live 98 in-game.
+- VDP2: Fix NBG per dot special priority calculations. Fixes priority issues in Mr. Bones. (#703)
 
 
 ## Version 0.2.1